refactored DeadlineTracker to Resources folder
