+++ backend/email-microservice/src/main/java/edu/oswego/cs/application/EmailApplication.java
@@ -1,7 +1,5 @@
 package edu.oswego.cs.application;
 
-import edu.oswego.cs.util.DeadlineTracker;
-
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
