+++ backend/email-microservice/src/main/java/edu/oswego/cs/resources/EmailResources.java
@@ -3,7 +3,6 @@ package edu.oswego.cs.resources;
 import edu.oswego.cs.database.AssignmentInterface;
 import edu.oswego.cs.database.CourseInterface;
 import edu.oswego.cs.services.EmailService;
-import edu.oswego.cs.util.DeadlineTracker;
 import org.bson.Document;
 
 import javax.annotation.security.DenyAll;
